ENH: Add intelligent selection of OpenMP and ITK threads to multiScaleLoG. Will require this ITK patch to work effectively: http://review.source.kitware.com/#/c/8769/

+++ NuclearSegmentation/yousef_core/seed_detection_3D/seedsdetection_3D.h
@@ -20,6 +20,6 @@
 
 // #include "itkMultiThreader.h"
 
-int Seeds_Detection_3D( float* IM, float** IM_out, unsigned short** IM_bin, int r, int c, int z, double *sigma_min, double *sigma_max, double *scale_xy, double *scale_z,int sampl_ratio,unsigned short* bImg, int UseDistMap, int* minIMout, bool paramEstimation);
+int Seeds_Detection_3D( float* IM, float** IM_out, unsigned short** IM_bin, size_t r, size_t c, size_t z, double *sigma_min, double *sigma_max, double *scale_xy, double *scale_z,int sampl_ratio, unsigned short* bImg, int UseDistMap, int* minIMout, bool paramEstimation);
 
 #endif
